pyproject.toml
[project]
name = "study-ocr"
version = "0.1.0"
description = "Hands-on OCR/VLM practice (LLaVA, BLIP-2 Q-Former, CLIP) with Jupyter"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate",
    "bitsandbytes",
    "jupyterlab",
    "ipykernel",
    "matplotlib",
    "Pillow",
    "sentencepiece",
    "torch",
    "torchvision",
    "transformers>=4.57.3",
    "ipywidgets",
    "pillow-heif",
    "torchinfo",
    "torchviz",
    "graphviz",
    "sentence-transformers",
    "sam3",
]
